+++ rainier-core/src/main/scala/com/stripe/rainier/ir/OutputMethodGenerator.scala
@@ -1,8 +1,8 @@
 package com.stripe.rainier.ir
 
-final private case class ApplyMethodGenerator(classPrefix: String,
-                                              classSizeLimit: Int,
-                                              outputMethods: Seq[Int])
+final private case class OutputMethodGenerator(classPrefix: String,
+                                               classSizeLimit: Int,
+                                               outputMethods: Seq[Int])
     extends MethodGenerator {
   val isStatic: Boolean = false
   val methodName: String = "output"
@@ -11,7 +11,7 @@ final private case class ApplyMethodGenerator(classPrefix: String,
   loadOutputIndex()
   tableSwitch(outputMethods) {
     case Some(m) => callExprMethod(m)
-    case None    => constant(0.0)
+    case None    => throwNPE() //easiest exception to generate
   }
   returnDouble()
 }
